add orcid-authorization header

+++ src/api/orcid/orcid.ts
@@ -1,6 +1,8 @@
 import api, { ADSQuery, ApiRequestConfig, ApiTargets } from '@api';
 import { QueryFunction, useQuery } from 'react-query';
 import { IOrcidParams, IOrcidResponse, IOrcidUser } from '@api/orcid/types';
+import { AppState, useStore } from '@store';
+import { isValidIOrcidUser } from '@api/orcid/models';
 
 export enum OrcidKeys {
   EXCHANGE_TOKEN = 'orcid/exchange-token',
@@ -33,11 +35,14 @@ export const useOrcidExchangeToken: OrcidQuery<'exchangeToken'> = (params, optio
   });
 };
 
+const orcidUserSelector = (store: AppState) => store.orcid.user;
 export const useOrcidGetProfile: OrcidQuery<'profile'> = (params, options) => {
+  const user = useStore(orcidUserSelector);
+
   return useQuery({
     queryKey: orcidKeys.profile(params),
     queryFn: fetchProfile,
-    meta: { params },
+    meta: { params, user },
     ...options,
   });
 };
@@ -91,13 +96,20 @@ const fetchExchangeToken: QueryFunction<IOrcidUser> = async ({ meta }) => {
 
 
 const fetchProfile: QueryFunction<IOrcidResponse['profile']> = async ({ meta }) => {
-  const { params } = meta as { params: IOrcidParams['profile'] };
+  const { params, user } = meta as { params: IOrcidParams['profile']; user: IOrcidUser };
+
+  if (!isValidIOrcidUser(user)) {
+    throw new Error('Invalid ORCiD User');
+  }
 
   const config: ApiRequestConfig = {
     method: 'GET',
     url: `${ApiTargets.ORCID}/${params.orcid}/${ApiTargets.ORCID_PROFILE}${params.full ?? true ? '/full' : ''}${
       params.update ?? true ? '?update=true' : ''
     }`,
+    headers: {
+      'orcid-authorization': `Bearer ${user.access_token}`,
+    },
   };
 
   const { data } = await api.request<IOrcidResponse['profile']>(config);
@@ -105,12 +117,19 @@ const fetchProfile: QueryFunction<IOrcidResponse['profile']> = async ({ meta }) 
 };
 
 const removeWorks: QueryFunction<IOrcidResponse['removeWorks']> = async ({ meta }) => {
-  const { params } = meta as { params: IOrcidParams['removeWorks'] };
+  const { params, user } = meta as { params: IOrcidParams['removeWorks']; user: IOrcidUser };
+
+  if (!isValidIOrcidUser(user)) {
+    throw new Error('Invalid ORCiD User');
+  }
 
   const config: ApiRequestConfig = {
     method: 'DELETE',
     // TODO handle bulk deletions
     url: `${ApiTargets.ORCID}/${params.orcid}/${ApiTargets.ORCID_WORKS}/${params.works[0]}`,
+    headers: {
+      'orcid-authorization': `Bearer ${user.access_token}`,
+    },
   };
 
   const { data } = await api.request<null>(config);
@@ -118,13 +137,20 @@ const removeWorks: QueryFunction<IOrcidResponse['removeWorks']> = async ({ meta 
 };
 
 const addWorks: QueryFunction<IOrcidResponse['addWorks']> = async ({ meta }) => {
-  const { params } = meta as { params: IOrcidParams['addWorks'] };
+  const { params, user } = meta as { params: IOrcidParams['addWorks']; user: IOrcidUser };
+
+  if (!isValidIOrcidUser(user)) {
+    throw new Error('Invalid ORCiD User');
+  }
 
   const config: ApiRequestConfig = {
     method: 'POST',
     url: `${ApiTargets.ORCID}/${params.orcid}/${ApiTargets.ORCID_WORKS}`,
     data: {
       bulk: params.works,
+    },
+    headers: {
+      'orcid-authorization': `Bearer ${user.access_token}`,
     },
   };
 
@@ -133,15 +159,28 @@ const addWorks: QueryFunction<IOrcidResponse['addWorks']> = async ({ meta }) => 
 };
 
 const updateWork: QueryFunction<IOrcidResponse['updateWork']> = async ({ meta }) => {
-  const { params } = meta as { params: IOrcidParams['updateWork'] };
+  const { params, user } = meta as { params: IOrcidParams['updateWork']; user: IOrcidUser };
+
+  if (!isValidIOrcidUser(user)) {
+    throw new Error('Invalid ORCiD User');
+  }
 
   const url = `${ApiTargets.ORCID}/${params.orcid}/${ApiTargets.ORCID_WORKS}/${params.work['put-code']}`;
 
-  const getConfig: ApiRequestConfig = { method: 'GET', url };
+  const getConfig: ApiRequestConfig = {
+    method: 'GET',
+    url,
+    headers: {
+      'orcid-authorization': `Bearer ${user.access_token}`,
+    },
+  };
   const putConfig: ApiRequestConfig = {
     method: 'PUT',
     url,
     data: params.work,
+    headers: {
+      'orcid-authorization': `Bearer ${user.access_token}`,
+    },
   };
 
   const res = await Promise.all([api.request<null>(getConfig), api.request<null>(putConfig)]);
@@ -151,11 +190,18 @@ const updateWork: QueryFunction<IOrcidResponse['updateWork']> = async ({ meta })
   return res[1].data;
 };
 const getName: QueryFunction<IOrcidResponse['name']> = async ({ meta }) => {
-  const { params } = meta as { params: IOrcidParams['name'] };
+  const { params, user } = meta as { params: IOrcidParams['name']; user: IOrcidUser };
+
+  if (!isValidIOrcidUser(user)) {
+    throw new Error('Invalid ORCiD User');
+  }
 
   const config: ApiRequestConfig = {
     method: 'GET',
     url: `${ApiTargets.ORCID_NAME}/${params.orcid}`,
+    headers: {
+      'orcid-authorization': `Bearer ${user.access_token}`,
+    },
   };
 
   const { data } = await api.request<null>(config);
